// src/common/enums/hazard.enum.ts
export enum HazardType {
  FIRE = 'FIRE',
  FLOOD = 'FLOOD',
  LANDSLIDE = 'LANDSLIDE',
  SINKHOLE = 'SINKHOLE',
  ROAD_DAMAGE = 'ROAD_DAMAGE',
  COLLAPSE = 'COLLAPSE',
  BUILDING_DAMAGE = 'BUILDING_DAMAGE',
  CHEMICAL = 'CHEMICAL',
  TRAFFIC = 'TRAFFIC',
  CONSTRUCTION = 'CONSTRUCTION',
  OTHER = 'OTHER',
}

export enum HazardLevel {
  LOW = 'LOW',
  MEDIUM = 'MEDIUM',
  HIGH = 'HIGH',
}